Finished Income page and added Budget page

+++ SmartSaver/DTO/Expenses/Output/SumByCatDTO.cs
@@ -6,7 +6,7 @@ namespace SmartSaver.DTO.Expenses.Input
 {
     class SumByCatDTO
     {
-        public double sum { get; set; }
+        public float sum { get; set; }
         public string category { get; set; }
 
     }
